refactor: move getPadding and getGridStyles to SafeAreaContext
